Moves files to /pkg
